CHROMIUM: drm/i915/pxp: Implement ioctl actions to manage a user session
This patch allows a user to reserve and manage a PXP session. Creating
a session requires a 3-way handshake between i915, userspace and CSME/GSC
firmware, which follows these steps:

1 - user reserves a session ID with i915.
2 - user starts the session with CSME/GSC.
3 - user notifies i915 that the session has been started.

Destroyng a session requires a submission to the HW, which is performed
by i915 upon receiving the ioctl call from the user.

All sessions are terminated and the relevand IDs released when a
termination interrupt is received.

v2: check against drm_file when setting session in play (Alan)

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed by: Alan Previn <alan.previn.teres.alexis@intel.com>
Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Acked-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Acked-by: Lucas De Marchi <lucas.demarchi@intel.com>

BUG=b:321196166
TEST=i915 driver loaded successfully
UPSTREAM-TASK=b:185364793

Change-Id: I71a2f95c736b0aa811ca6ae6820a5f201a368cd8
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5231520
Tested-by: Altaf Basha <altaf.basha@intel.com>
Reviewed-by: Juston Li <justonli@google.com>