FROMLIST: drm/vkms: Add ConfigFS scaffolding to VKMS
This change adds the basic scaffolding for ConfigFS, including setting
up the default directories. It does not allow for the registration of
configfs-backed devices, which is complex and provided in a follow-up
commit.

This CL includes docs about using ConfigFS with VKMS, but I'll summarize
in brief here as well (assuming ConfigFS is mounted at /config/):

To create a new device, you can do so via `mkdir
/config/vkms/my-device`.

This will create a number of directories and files automatically:

	/config
	`-- vkms
	    `-- my-device
		|-- connectors
		|-- crtcs
		|-- encoders
		|-- planes
		`-- enabled

You can then configure objects by mkdir'ing in each of the directories.

When you're satisfied, you can `echo 1 > /config/vkms/my-device/enabled`.
This will create a new device according to your configuration.

For now, this will fail, but the next change will add support for it.

Signed-off-by: Jim Shargo <jshargo@chromium.org>
Signed-off-by: Brandon Pollack <brpol@chromium.org>
(am from https://patchwork.kernel.org/project/dri-devel/patch/20230829053201.423261-5-brpol@chromium.org/)

The change is in the final stages of review (fifth iteration), All that
remains is some (potential) igt tests and final signoff.

Some reasons we'd like to submit in chromeOS are:

This will allow our virtual multi display tests to run in presubmit.
This real world case will be compelling evidence if/when upstream
requests some data that this ran for real in production (and avoiding
the chicken egg problem of we need this in the kernel to run the tests
but without running the tests we can't submit upstream).

Signed-off-by: Brandon Pollack <brpol@chromium.org>
BUG=241800059
BUG=283357160
TEST=new tests defined here passing https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4666669

Change-Id: I853254cbcb439ee58d6e538b51fab8a2aa842ca3
Signed-off-by: Brandon Pollack <brpol@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4810101
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>